+++ src/uk/ac/tvu/mdse/contextengine/highLevelContext/ContextRange.java
@@ -23,18 +23,18 @@ public class ContextRange {
 	public static final String LOG_TAG = "ContextRange";
 	public static final boolean D = true;
 
-	public int minValue;
-	public int maxValue;
+	public long minValue;
+	public long maxValue;
 	public String contextHighValue;
 	
-	public ContextRange(int mv, int mxv, String value){
+	public ContextRange(long mv, long mxv, String value){
 		if (D) Log.d(LOG_TAG, "constructor");
 		this.minValue = mv;
 		this.maxValue = mxv;
 		this.contextHighValue = value;
 	}
 	
-	public String getContextHighValue(int currentValue){
+	public String getContextHighValue(long currentValue){
 		if (D) Log.d(LOG_TAG, "getContextHighValue");
 		return ((currentValue<maxValue)&&(currentValue>=minValue)) ? contextHighValue : null;
 	}
